packages/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "minions-clients"
version = "0.1.0"
description = "Client profiles built from job interactions, communications, and contracts"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
keywords = ["clients","ai","minions"]
authors = [
    { name = "Mehdi Nabhani", email = "mehdi@the-mehdi.com" }
]
dependencies = [
    "minions-sdk>=0.2.1",
]

[project.optional-dependencies]
test = ["pytest>=7.0"]

[project.urls]
Homepage = "https://github.com/mxn2020/minions-clients"
Repository = "https://github.com/mxn2020/minions-clients"

[tool.hatch.build.targets.wheel]
packages = ["minions_clients"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
